feat: add user and model management
Design: REQ-005, REQ-006, REQ-017, UI-009, UI-012, HD-014, HD-019, ADR-026

Tests: Provider 8/8, Core 24/24, WebApi 18/18, Electron E2E 3/3

Verify: solution and AppHost builds, desktop build/lint, staged diff check

Docs: update H5-008 and H5-013 implementation records

Risk: authenticated live WebApi smoke and non-Chat diagnostics remain

Task: implement shared user/model list workflows and LiteLLM management metadata

+++ src/core/providers/LLM/Inkwell.LLM.LiteLLM/LiteLLMModelGroupDataItem.cs
@@ -25,13 +25,13 @@ internal sealed class LiteLLMModelGroupDataItem
     /// 获取最大输入 token 数。
     /// </summary>
     [JsonPropertyName("max_input_tokens")]
-    public int? MaxInputTokens { get; init; }
+    public double? MaxInputTokens { get; init; }
 
     /// <summary>
     /// 获取最大输出 token 数。
     /// </summary>
     [JsonPropertyName("max_output_tokens")]
-    public int? MaxOutputTokens { get; init; }
+    public double? MaxOutputTokens { get; init; }
 
     /// <summary>
     /// 获取是否支持函数调用。
